New forked deps concept
Use wheels in sys.paths to make libpdf aware of them.
The wheels are delivered in the deps folder which is
included in pyproject.toml.
The new file _import_forks.py checks whether the forks
are available for import from the environment. If not
the wheels from the deps folder are added to sys.path.